+++ EddiCore/ISpeechResponderController.cs
@@ -8,6 +8,8 @@ namespace EddiCore
     {
         Task SayAsync ( Ship ship, string scriptName, Event theEvent = null, int? priority = null, string voice = null, bool sayOutLoud = true, bool invokedFromVA = false );
 
+        string Humanise ( decimal? number );
+
         bool TrySetPersonality ( string newPersonalityName );
     }
 }