+++ ScriptResolverService/CustomFunctions/Humanise.cs
@@ -1,17 +1,23 @@
 using Cottle;
-using EddiSpeechService.SpeechConversions;
 using JetBrains.Annotations;
 using System;
+using System.Collections.Generic;
 
 namespace EddiScriptResolverService.CustomFunctions
 {
     [UsedImplicitly]
-    public class Humanise : ICustomFunction
+    [ method: UsedImplicitly]
+    public class Humanise ( IContext context, IReadOnlyDictionary<string, IScriptDefinition> scripts )
+        : RecursiveFunction( context, scripts ), ICustomFunction
     {
         public string name => "Humanise";
         public FunctionCategory Category => FunctionCategory.Utility;
         public string description => Properties.CustomFunctions_Untranslated.Humanise;
         public Type ReturnType => typeof( string );
-        public IFunction function => Function.CreateNative1( ( runtime, input, writer ) => SpeechConversions.Humanize( (decimal?)Convert.ToDecimal( input.AsNumber ) ) );
+        public IFunction function => Function.CreateNative1( ( runtime, input, writer ) =>
+        {
+            var number = (decimal?)Convert.ToDecimal( input.AsNumber );
+            return HumaniseRenderer.Render( number, Scripts, GetContext( runtime.Globals ) );
+        } );
     }
 }